Move Nested logic to NestedTransformer
