+++ src/main/java/com/brunodles/jsoupparser/smallanotation/SmallAnnotationInvocationHandler.java
@@ -9,11 +9,12 @@ import com.brunodles.jsoupparser.smallanotation.collectors.AttrCollector;
 import com.brunodles.jsoupparser.smallanotation.collectors.AttrCollectorTransformer;
 import com.brunodles.jsoupparser.smallanotation.collectors.TextCollector;
 import com.brunodles.jsoupparser.smallanotation.collectors.TextCollectorTransformer;
+import com.brunodles.jsoupparser.smallanotation.nested.Nested;
+import com.brunodles.jsoupparser.smallanotation.nested.NestedTransformer;
 import com.brunodles.jsoupparser.smallanotation.selector.Selector;
 import com.brunodles.jsoupparser.smallanotation.selector.SelectorTransformer;
 import com.brunodles.jsoupparser.smallanotation.withtype.WithType;
 import com.brunodles.jsoupparser.smallanotation.withtype.WihTypeTransformer;
-import org.jsoup.nodes.Element;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.InvocationHandler;
@@ -31,6 +32,7 @@ public class SmallAnnotationInvocationHandler implements MethodInvocationHandler
         transfomers.put(TextCollector.class, TextCollectorTransformer.class);
         transfomers.put(AttrCollector.class, AttrCollectorTransformer.class);
         transfomers.put(WithType.class, WihTypeTransformer.class);
+        transfomers.put(Nested.class, NestedTransformer.class);
         transformerMap = Collections.unmodifiableMap(transfomers);
     }
 
@@ -58,13 +60,6 @@ public class SmallAnnotationInvocationHandler implements MethodInvocationHandler
                 } catch (InstantiationException | IllegalAccessException e) {
                     throw new RuntimeException(e);
                 }
-                continue;
-            }
-            if (annotation instanceof NestedCollector && result != null) {
-                List newResult = new ArrayList<>(result.size());
-                for (Object o : result)
-                    newResult.add(invocation.proxyHandler.jsoupParser.parseElement(((Element) o), invocation.getMethodRealReturnType()));
-                result = newResult;
                 continue;
             }
             if (annotation instanceof FollowTransformer && result != null && !result.isEmpty()) {
