+++ src/main/java/com/brunodles/jsoupparser/smallanotation/nested/Nested.java
@@ -1,4 +1,4 @@
-package com.brunodles.jsoupparser.smallanotation.nestedcollector;
+package com.brunodles.jsoupparser.smallanotation.nested;
 
 import java.lang.annotation.*;
